requirements.txt
spacy==3.7.5
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
python-dateutil==2.9.0.post0
colorama==0.4.6
